Allow HTTPS for GET requests
We found some service that required an HTTPS call on a GET. Typically this means our elegant API is slightly spoilt.

I added an additional function to allow the scheme to be set because typically you don't need anything but HTTP for GET OAuth requests and well designed APIs should allow you to use 80% of their feature sets without bloat.